pyproject.toml
[tool.poetry]
name = "pistacchio"
version = "0.1.3-10"
description = "Yet another Federated Learning framework"
authors = ["Luca Corbucci"]
license = "MIT"
readme = "README.md"
packages = [ { include = "pistacchio" } ]

[tool.poetry.dependencies]
python = "^3.10"
torch = "^1.13.1"
numpy = "^1.24.1"
mypy = "^0.991"
opacus = "^1.3.0"
scikit-learn = "^1.2.1"
torchsummary = "^1.5.1"
tqdm = "^4.64.1"
python-dotenv = "^0.21.1"
regex = "^2022.10.31"
wandb = "^0.13.9"
pandas = "^1.5.3"
dill = "^0.3.6"
multiprocess = "^0.70.14"
torchvision = "^0.14.1"
pydantic = "^1.10.4"
ruff = "^0.0.238"
pydoc-markdown = "^4.6.4"
loguru = "^0.6.0"

[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"
black = {extras = ["jupyter"], version = "^22.12.0"}
pytest = "^7.2.1"
coverage = "^7.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# --- Black config ------------------------------------------- #

[tool.black]
line-length = 88

# --- isort config ------------------------------------------- #

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 88
lines_after_imports = 2



[tool.ruff]
line-length = 88
select = ["ALL"]
target-version = "py310"
ignore = [
    "D101",
    "D100",
    "D404",
    "I001",
    "C901",
    "D213",
    "D203",
    "B905",
    "EM101",
    "D205",
    "D407",
    "E501",
    "ANN101",
    "D401",
    "T201",
    "FBT001",
    "B006",
    "PLR0913", # too many arguments in function
    "G004",
    "S101",
]
fix = true
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    "tests/codegen/snapshots"
]
src = ["pistacchio", "tests"]

[tool.ruff.isort]
known-first-party = ["pistacchio"]
extra-standard-library = ["typing_extensions"]

